+++ sdk/src/beta9/__init__.py
@@ -1,6 +1,6 @@
 from . import env
 from .abstractions import experimental, integrations
-from .abstractions.container import Container
+from .abstractions.base.container import Container
 from .abstractions.endpoint import ASGI as asgi
 from .abstractions.endpoint import Endpoint as endpoint
 from .abstractions.endpoint import RealtimeASGI as realtime
