+++ sdk/src/beta9/cli/container.py
@@ -1,27 +1,19 @@
 import datetime
-import os
-import time
-from queue import Empty, Queue
-from typing import Callable, List, Optional
+from typing import List
 
 import click
 from betterproto import Casing
 from rich.table import Column, Table, box
-from watchdog.observers import Observer
 
 from .. import terminal
+from ..abstractions.base.container import Container
 from ..channel import ServiceClient
 from ..cli import extraclick
 from ..clients.gateway import (
-    AttachToContainerRequest,
-    ContainerStreamMessage,
     ListContainersRequest,
     StopContainerRequest,
     StopContainerResponse,
-    SyncContainerWorkspaceOperation,
-    SyncContainerWorkspaceRequest,
 )
-from ..sync import SyncEventHandler
 from .extraclick import ClickCommonGroup, ClickManagementGroup
 
 
@@ -153,79 +145,6 @@ def stop_container(service: ServiceClient, container_ids: List[str]):
             terminal.error(f"{res.error_msg}", exit=False)
 
 
-def sync_dir_to_workspace(*, dir: str, container_id: str, on_event: Optional[Callable] = None):
-    file_update_queue = Queue()
-    event_handler = SyncEventHandler(file_update_queue)
-
-    observer = Observer()
-    observer.schedule(event_handler, dir, recursive=True)
-    observer.start()
-
-    terminal.header(f"Watching '{dir}' for changes...")
-    while True:
-        try:
-            operation, path, new_path = file_update_queue.get_nowait()
-
-            if on_event:
-                on_event(operation, path, new_path)
-
-            req = SyncContainerWorkspaceRequest(
-                container_id=container_id,
-                path=os.path.relpath(path, start=dir),
-                is_dir=os.path.isdir(path),
-                op=operation,
-            )
-
-            if operation == SyncContainerWorkspaceOperation.WRITE:
-                if not req.is_dir:
-                    with open(path, "rb") as f:
-                        req.data = f.read()
-
-            elif operation == SyncContainerWorkspaceOperation.DELETE:
-                pass
-
-            elif operation == SyncContainerWorkspaceOperation.MOVED:
-                req.new_path = os.path.relpath(new_path, start=dir)
-
-            yield ContainerStreamMessage(sync_container_workspace=req)
-
-            file_update_queue.task_done()
-        except Empty:
-            time.sleep(0.1)
-        except Exception as e:
-            terminal.warn(str(e))
-
-
-def _attach_to_container(service: ServiceClient, container_id: str):
-    terminal.header(f"Connecting to {container_id}...")
-
-    def _container_stream_generator():
-        yield ContainerStreamMessage(
-            attach_request=AttachToContainerRequest(container_id=container_id)
-        )
-
-        yield from sync_dir_to_workspace(dir="./", container_id=container_id)
-
-    # Connect to the remote container and stream messages back and forth
-    stream = service.gateway.attach_to_container(_container_stream_generator())
-
-    r = None
-    for r in stream:
-        if r.output:
-            terminal.detail(r.output, end="")
-
-        if r.done or r.exit_code != 0:
-            break
-
-    if r is None:
-        return terminal.error("Container failed ❌")
-
-    if not r.done or r.exit_code != 0:
-        return terminal.error(f"\n{r.output} ❌")
-
-    terminal.success(r.output)
-
-
 @management.command(
     name="attach",
     help="Attach to a running container.",
@@ -236,4 +155,5 @@ def _attach_to_container(service: ServiceClient, container_id: str):
 )
 @extraclick.pass_service_client
 def attach_to_container(service: ServiceClient, container_id: str):
-    _attach_to_container(service, container_id)
+    container = Container(container_id=container_id)
+    container.attach(service, container_id)
